+++ src/main/java/app/Main.java
@@ -1,6 +1,6 @@
 package app;
 
-import app.api.ApiReader;
+import app.services.ApiReader;
 
 public class Main {
     public static void main(String[] args) {
@@ -8,8 +8,8 @@ public class Main {
         ApiReader apiReader = new ApiReader();
 
         apiReader.readMovieMultiple();
-        apiReader.apiMovies.forEach(System.out::println);
-        apiReader.apiCredits.forEach(System.out::println);
+//        apiReader.apiMovies.forEach(System.out::println);
+//        apiReader.apiCredits.forEach(System.out::println);
         System.out.println("Amount of movies: " + apiReader.apiMovies.size());
         System.out.println("Amount of credits: " + apiReader.apiCredits.size());
