+++ src/main/java/app/entities/special_entities/MovieCredits.java
@@ -1,7 +1,7 @@
 package app.entities.special_entities;
 
-import app.api.custom_deserializers.CastIdDeserializer;
-import app.api.custom_deserializers.CrewIdDeserializer;
+import app.custom_deserializers.CastIdDeserializer;
+import app.custom_deserializers.CrewIdDeserializer;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
