+++ src/main/java/app/entities/special_entities/MovieMulti.java
@@ -1,6 +1,6 @@
 package app.entities.special_entities;
 
-import app.api.custom_deserializers.MovieMultiIdDeserializer;
+import app.custom_deserializers.MovieMultiIdDeserializer;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
